requirements.txt
PySide6
pyserial
Cartopy
matplotlib